[command,account,keypair] refactor seed generation to account
remove obsolete keypair
