+++ account/private.go
@@ -7,11 +7,8 @@ package account
 
 import (
 	"bytes"
-	"crypto/rand"
-	"fmt"
 
 	"golang.org/x/crypto/ed25519"
-	"golang.org/x/crypto/nacl/secretbox"
 	"golang.org/x/crypto/sha3"
 
 	"github.com/bitmark-inc/bitmarkd/fault"
@@ -44,128 +41,6 @@ type ED25519PrivateKey struct {
 type NothingPrivateKey struct {
 	Test       bool
 	PrivateKey []byte
-}
-
-// seed parameters
-var (
-	seedHeader   = []byte{0x5a, 0xfe}
-	seedHeaderV1 = append(seedHeader, []byte{0x01}...)
-	seedHeaderV2 = append(seedHeader, []byte{0x02}...)
-)
-
-// for seed v1 only
-var (
-	seedNonce = [24]byte{
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	}
-	authSeedIndex = [16]byte{
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0xe7,
-	}
-)
-
-const (
-	seedHeaderLength   = 3
-	seedPrefixLength   = 1
-	secretKeyV1Length  = 32
-	secretKeyV2Length  = 17
-	seedChecksumLength = 4
-
-	seedV1Length = 40
-	seedV2Length = 24
-)
-
-// PrivateKeyFromBase58Seed - this converts a Base58 encoded seed string and returns a private key
-//
-// one of the specific private key types are returned using the base "PrivateKeyInterface"
-// interface type to allow individual methods to be called.
-func PrivateKeyFromBase58Seed(seedBase58Encoded string) (*PrivateKey, error) {
-
-	// verify length
-	seed := util.FromBase58(seedBase58Encoded)
-	seedLength := len(seed)
-	if seedV1Length != seedLength && seedV2Length != seedLength {
-		return nil, fault.ErrInvalidSeedLength
-	}
-
-	// verify checksum
-	digest := sha3.Sum256(seed[:seedLength-checksumLength])
-	checksumStart := seedLength - seedChecksumLength
-	expectedChecksum := digest[:seedChecksumLength]
-	actualChecksum := seed[checksumStart:]
-	if !bytes.Equal(expectedChecksum, actualChecksum) {
-		return nil, fault.ErrChecksumMismatch
-	}
-
-	header := seed[:seedHeaderLength]
-	var ed25519Seed []byte // the ed25519 seed for generate key pair
-	var testnet bool       // denote the network is test net
-
-	switch {
-	case bytes.Equal(seedHeaderV1, header):
-		// copy the secret key from seed
-		var sk [secretKeyV1Length]byte
-		secretStart := seedHeaderLength + seedPrefixLength
-		copy(sk[:], seed[secretStart:])
-
-		prefix := seed[seedHeaderLength:secretStart]
-		// first byte of prefix is test/live indication
-		testnet = prefix[0] == 0x01
-
-		ed25519Seed = secretbox.Seal([]byte{}, authSeedIndex[:], &seedNonce, &sk)
-
-	case bytes.Equal(seedHeaderV2, header):
-		sk := seed[seedHeaderLength:checksumStart]
-
-		// verify valid secret key
-		if secretKeyV2Length != len(sk) || 0 != sk[16]&0x0f {
-			return nil, fault.ErrInvalidSeedLength
-		}
-
-		// parse network
-		mode := sk[0]&0x80 | sk[1]&0x40 | sk[2]&0x20 | sk[3]&0x10
-		testnet = mode == sk[15]&0xf0^0xf0
-
-		// add the seed 4 times to hash value
-		hash := sha3.NewShake256()
-		for i := 0; i < 4; i++ {
-			n, err := hash.Write(sk)
-			if err != nil {
-				return nil, err
-			}
-			if secretKeyV2Length != n {
-				return nil, fault.ErrCannotDecodeSeed
-			}
-		}
-
-		ed25519Seed = make([]byte, ed25519.SeedSize)
-		n, err := hash.Read(ed25519Seed)
-		if nil != err {
-			return nil, err
-		}
-		if ed25519.SeedSize != n {
-			return nil, fault.ErrCannotDecodeSeed
-		}
-
-	default:
-		return nil, fault.ErrInvalidSeedHeader
-	}
-
-	// generate key pair from encrypted secret key
-	_, priv, err := ed25519.GenerateKey(bytes.NewBuffer(ed25519Seed))
-	if nil != err {
-		return nil, err
-	}
-
-	privateKey := &PrivateKey{
-		PrivateKeyInterface: &ED25519PrivateKey{
-			Test:       testnet,
-			PrivateKey: priv,
-		},
-	}
-	return privateKey, nil
 }
 
 // PrivateKeyFromBase58 - this converts a Base58 encoded string and returns an private key
@@ -403,45 +278,4 @@ func (privateKey *NothingPrivateKey) String() string {
 // MarshalText - convert an privateKey to its Base58 JSON form
 func (privateKey NothingPrivateKey) MarshalText() ([]byte, error) {
 	return []byte(privateKey.String()), nil
-}
-
-// GenerateEncodedSeedV2 generate base58 seed v2
-func GenerateEncodedSeedV2(testnet bool) (string, error) {
-
-	// space for 128 bit, extend to 132 bit later
-	sk := make([]byte, 16, secretKeyV2Length)
-
-	n, err := rand.Read(sk)
-	if nil != err {
-		return "", err
-	}
-
-	if 16 != n {
-		return "", fmt.Errorf("got %d bytes, expected is 16 bytes", n)
-	}
-
-	// extend to 132 bits
-	sk = append(sk, sk[15]&0xf0)
-
-	if secretKeyV2Length != len(sk) {
-		return "", fmt.Errorf("actual seed length is %d bytes, expected is %d bytes", len(sk), secretKeyV2Length)
-	}
-
-	// network flag
-	mode := sk[0]&0x80 | sk[1]&0x40 | sk[2]&0x20 | sk[3]&0x10
-	if testnet {
-		mode = mode ^ 0xf0
-	}
-	sk[15] = mode | sk[15]&0x0f
-
-	// encode seed to base58
-	seed := make([]byte, 0)
-	seed = append(seed, seedHeaderV2...)
-	seed = append(seed, sk...)
-	digest := sha3.Sum256(seed)
-	checksum := digest[:seedChecksumLength]
-	seed = append(seed, checksum...)
-	result := util.ToBase58(seed)
-
-	return result, nil
 }
