--- keypair/doc.go
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: ISC
-// Copyright (c) 2014-2019 Bitmark Inc.
-// Use of this source code is governed by an ISC
-// license that can be found in the LICENSE file.
-
-// Package keypair - Bitmark keypair
-package keypair
